pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prompt-library"
version = "0.1.0"
description = "Enterprise prompt engineering library with testing and management"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Onurcan Dönmezer", email = "onurcan.donmezer@gmail.com"},
]
keywords = ["llm", "prompt-engineering", "gemini", "ai", "enterprise"]

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "google-genai>=1.0",
    "streamlit>=1.30",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.3",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
